Update User Class, Service and Repository

+++ src/main/java/org/Aruiz/repository/UserRepository.java
@@ -9,13 +9,13 @@ import java.util.Optional;
 
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
     // Método para buscar un usuario por su nombre de usuario (único)
     Optional<User> findByUsername(String username);
 
     // Método para eliminar un usuario por su nombre de usuario (único)
 
-    void deleteById(Long id);
+    void deleteById(int id);
 
 }
